+++ ScarleCodeBase-master/DXTK_Wrapper/DXTKRenderer.cpp
@@ -97,14 +97,14 @@ float Renderer::getWindowWidth()
 {
 	RECT rc;
 	GetClientRect(hWnd, &rc);
-	return rc.right - rc.left;
+	return (float)(rc.right - rc.left);
 }
 
 float Renderer::getWindowHeight()
 {
 	RECT rc;
 	GetClientRect(hWnd, &rc);
-	return rc.bottom - rc.top;
+	return (float)(rc.bottom - rc.top);
 }
 
 //void Renderer::DrawString(wchar_t const * text, Vec2 const & position, Vec4 const & color, float rotation, Vec2 const & origin, Vec2 const & scale, float layerDepth) const
