+++ ScarleCodeBase-master/Main_Engine/Background.cpp
@@ -13,7 +13,7 @@ Background::Background(Sprite * _sprite, BaseCamera * camera)
 	sprite = _sprite;
 	parallax = false;
 	sprite->setPosition(main_camera->getPosition() * -1);
-	Vec2 screenSize(GameDataV2::screen_width, GameDataV2::screen_height);
+	Vec2 screenSize((float)GameDataV2::screen_width, (float)GameDataV2::screen_height);
 	Vec2 scaleFactor(screenSize.x / sprite->getSize().x, screenSize.y / sprite->getSize().y);
 	sprite->setScale(scaleFactor);
 }
@@ -21,7 +21,7 @@ Background::Background(Sprite * _sprite, BaseCamera * camera)
 Background::Background(std::vector<Sprite*> _sprites, BaseCamera * camera)
 	: main_camera(camera)
 {
-	for (int i = 0; i < _sprites.size(); i++)
+	for (unsigned int i = 0; i < _sprites.size(); i++)
 	{
 		_sprites[i]->setOrigin(_sprites[i]->GetTexture()->getSize() * 0.5f);
 		backgrounds.push_back(new Sprite(*_sprites[i]));
@@ -29,9 +29,9 @@ Background::Background(std::vector<Sprite*> _sprites, BaseCamera * camera)
 		backgrounds.push_back(new Sprite(*_sprites[i]));
 		
 	}
-	for (int i = 0; i < backgrounds.size(); i+=3)
+	for (unsigned int i = 0; i < backgrounds.size(); i+=3)
 	{
-		Vec2 screenSize(GameDataV2::screen_width, GameDataV2::screen_height);
+		Vec2 screenSize((float)GameDataV2::screen_width, (float)GameDataV2::screen_height);
 		Vec2 scaleFactor(2, 2);/*screenSize.x / backgrounds[i]->getSize().x,
 						 screenSize.y / backgrounds[i]->getSize().y);*/
 		backgrounds[i]->setScale(scaleFactor /* (backgrounds.size() - i)*/);
@@ -43,7 +43,7 @@ Background::Background(std::vector<Sprite*> _sprites, BaseCamera * camera)
 		backgrounds[i]->setPosition(centerPosition);
 		backgrounds[i+1]->setPosition(Vec2(centerPosition.x - (backgrounds[i]->getSize().x * backgrounds[i]->getScale().x), centerPosition.y));
 		backgrounds[i+2]->setPosition(Vec2(centerPosition.x + (backgrounds[i]->getSize().x * backgrounds[i]->getScale().x), centerPosition.y));
-		parallaxScales.push_back(-1 * i);
+		parallaxScales.push_back(-1.0f * i);
 	}
 	parallax = true;
 	previousCamPosition += main_camera->getPosition() *-1;
@@ -57,7 +57,7 @@ bool Background::Update()
 {
 	if (parallax)
 	{
-		for (int i = 0; i < backgrounds.size(); i+=3)
+		for (unsigned int i = 0; i < backgrounds.size(); i+=3)
 		{
 			float parallax = previousCamPosition.x - (main_camera->getPosition().x );
 			Vec2 backgroundTargetPosition;
@@ -86,7 +86,7 @@ bool Background::Draw()
 {
 	if (parallax)
 	{
-		for (int i = 0; i < backgrounds.size(); i+=3)
+		for (unsigned int i = 0; i < backgrounds.size(); i+=3)
 		{
 			Sprite* tmp = backgrounds[i];
